Time: 0 ms (100%), Space: 8.6 MB (51.12%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     int mySqrt(int x) {
-        int left = 0;
-        int right = x;
-        int sol = 0;
+        long long left = 0;
+        long long right = x;
+        long long sol = 0;
 
         while (left <= right)
         {
